requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.40.0
yfinance
pandas==2.2.0
requests==2.32.0
pydantic==2.9.0
python-dotenv==1.0.0
